+++ tests/Unit/Public/Update-JaCoCoStatistic.tests.ps1
@@ -8,71 +8,7 @@ Import-Module $ProjectName
 
 Describe 'Update-JaCoCoStatistic' {
     BeforeAll {
-        function Get-XmlAttributeValue
-        {
-            [CmdletBinding()]
-            [OutputType([Hashtable])]
-            param
-            (
-                [Parameter(Mandatory = $true, ValueFromPipeline = $true)]
-                [System.Xml.XmlDocument]
-                $XmlDocument,
-
-                [Parameter(Mandatory = $true)]
-                [System.String]
-                $XPath,
-
-                [Parameter()]
-                [Switch]
-                $LineAttributes
-            )
-
-            $counter = $XmlDocument | Select-XML -XPath $XPath
-
-            if ($LineAttributes.IsPresent)
-            {
-                $counterValues = @{
-                    ci = ($counter.Node | Select-XML -XPath '@ci').Node.Value
-                    mi = ($counter.Node | Select-XML -XPath '@mi').Node.Value
-                    cb = ($counter.Node | Select-XML -XPath '@cb').Node.Value
-                    mb = ($counter.Node | Select-XML -XPath '@mb').Node.Value
-                }
-            }
-            else
-            {
-                $counterValues = @{
-                    Missed = ($counter.Node | Select-XML -XPath '@missed').Node.Value
-                    Covered = ($counter.Node | Select-XML -XPath '@covered').Node.Value
-                }
-            }
-
-            return $counterValues
-        }
-
-        function Write-IndentedXml
-        {
-            [CmdletBinding()]
-            param
-            (
-                [Parameter(Mandatory = $true)]
-                [System.Xml.XmlDocument]
-                $XmlDocument
-            )
-
-            $StringWriter = New-Object -TypeName 'System.IO.StringWriter'
-            $XmlWriter = New-Object -TypeName 'System.XMl.XmlTextWriter' -ArgumentList $StringWriter
-
-            $xmlWriter.Formatting = 'indented'
-            $xmlWriter.Indentation = 2
-
-            $XmlDocument.WriteContentTo($XmlWriter)
-
-            $XmlWriter.Flush()
-
-            $StringWriter.Flush()
-
-            Write-Verbose -Message $StringWriter.ToString() -Verbose
-        }
+        . $PSScriptRoot/../TestHelpers/Get-XmlAttribute.ps1
 
         $mockXmlDocument = @"
 <?xml version="1.0" encoding="us-ascii" standalone="no"?>
@@ -130,237 +66,233 @@ Describe 'Update-JaCoCoStatistic' {
                 $Line
             )
 
-            $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+            $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
             $result | Should -BeOfType [System.Xml.XmlDocument]
 
             $xPath = '/report/package[@name="3.0.0"]/sourcefile[@name="prefix.ps1"]/line[@nr="{0}"]' -f $Line
 
-            ($result | Get-XmlAttributeValue -XPath $xPath -LineAttributes).mi | Should -Be 0
-            ($result | Get-XmlAttributeValue -XPath $xPath -LineAttributes).ci | Should -Be 1
-            ($result | Get-XmlAttributeValue -XPath $xPath -LineAttributes).mb | Should -Be 0
-            ($result | Get-XmlAttributeValue -XPath $xPath -LineAttributes).cb | Should -Be 0
-
-            # Write-IndentedXml $result
+            ($result | Get-XmlAttribute -XPath $xPath).mi | Should -Be 0
+            ($result | Get-XmlAttribute -XPath $xPath).ci | Should -Be 1
+            ($result | Get-XmlAttribute -XPath $xPath).mb | Should -Be 0
+            ($result | Get-XmlAttribute -XPath $xPath).cb | Should -Be 0
         }
 
         Context 'When a package exist' {
             It 'Should return the correct statistics for the package counter INSTRUCTION' {
-                $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                 $result | Should -BeOfType [System.Xml.XmlDocument]
 
                 $xPath = '/report/package[@name="3.0.0"]/counter[@type="INSTRUCTION"]'
 
-                ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
+                ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
             }
 
             It 'Should return the correct statistics for the package counter LINE' {
-                $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                 $result | Should -BeOfType [System.Xml.XmlDocument]
 
                 $xPath = '/report/package[@name="3.0.0"]/counter[@type="LINE"]'
 
-                ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
+                ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
             }
 
             It 'Should return the correct statistics for the package counter METHOD' {
-                $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                 $result | Should -BeOfType [System.Xml.XmlDocument]
 
                 $xPath = '/report/package[@name="3.0.0"]/counter[@type="METHOD"]'
 
-                ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 1
+                ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 1
             }
 
             It 'Should return the correct statistics for the package counter CLASS' {
-                $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                 $result | Should -BeOfType [System.Xml.XmlDocument]
 
                 $xPath = '/report/package[@name="3.0.0"]/counter[@type="CLASS"]'
 
-                ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 1
+                ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 1
             }
 
             Context 'When sourcefile for class exist' {
                 It 'Should return the correct statistics for the sourcefile counter INSTRUCTION' {
-                    $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                    $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                     $result | Should -BeOfType [System.Xml.XmlDocument]
 
                     $xPath = '/report/package[@name="3.0.0"]/sourcefile[@name="prefix.ps1"]/counter[@type="INSTRUCTION"]'
 
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
+                    ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                    ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
                 }
 
                 It 'Should return the correct statistics for the sourcefile counter LINE' {
-                    $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                    $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                     $result | Should -BeOfType [System.Xml.XmlDocument]
 
                     $xPath = '/report/package[@name="3.0.0"]/sourcefile[@name="prefix.ps1"]/counter[@type="LINE"]'
 
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
+                    ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                    ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
                 }
 
                 It 'Should return the correct statistics for the sourcefile counter METHOD' {
-                    $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                    $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                     $result | Should -BeOfType [System.Xml.XmlDocument]
 
                     $xPath = '/report/package[@name="3.0.0"]/sourcefile[@name="prefix.ps1"]/counter[@type="METHOD"]'
 
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 1
+                    ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                    ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 1
                 }
 
                 It 'Should return the correct statistics for the sourcefile counter CLASS' {
-                    $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                    $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                     $result | Should -BeOfType [System.Xml.XmlDocument]
 
                     $xPath = '/report/package[@name="3.0.0"]/sourcefile[@name="prefix.ps1"]/counter[@type="CLASS"]'
 
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 1
+                    ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                    ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 1
                 }
             }
 
             Context 'When the package contain a class' {
                 It 'Should return the correct statistics for the class counter INSTRUCTION' {
-                    $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                    $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                     $result | Should -BeOfType [System.Xml.XmlDocument]
 
                     $xPath = '/report/package[@name="3.0.0"]/class/counter[@type="INSTRUCTION"]'
 
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
+                    ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                    ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
                 }
 
                 It 'Should return the correct statistics for the class counter LINE' {
-                    $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                    $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                     $result | Should -BeOfType [System.Xml.XmlDocument]
 
                     $xPath = '/report/package[@name="3.0.0"]/class/counter[@type="LINE"]'
 
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
+                    ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                    ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
                 }
 
                 It 'Should return the correct statistics for the class counter METHOD' {
-                    $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                    $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                     $result | Should -BeOfType [System.Xml.XmlDocument]
 
                     $xPath = '/report/package[@name="3.0.0"]/class/counter[@type="METHOD"]'
 
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 1
+                    ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                    ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 1
                 }
 
                 It 'Should return the correct statistics for the class counter CLASS' {
-                    $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                    $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                     $result | Should -BeOfType [System.Xml.XmlDocument]
 
                     $xPath = '/report/package[@name="3.0.0"]/class/counter[@type="CLASS"]'
 
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                    ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 1
+                    ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                    ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 1
                 }
 
                 Context 'When the class contain a method' {
                     It 'Should return the correct statistics for the method counter INSTRUCTION' {
-                        $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                        $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                         $result | Should -BeOfType [System.Xml.XmlDocument]
 
                         $xPath = '/report/package[@name="3.0.0"]/class/method[@name="<script>"]/counter[@type="INSTRUCTION"]'
 
-                        ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                        ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
+                        ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                        ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
                     }
 
                     It 'Should return the correct statistics for the method counter LINE' {
-                        $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                        $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                         $result | Should -BeOfType [System.Xml.XmlDocument]
 
                         $xPath = '/report/package[@name="3.0.0"]/class/method[@name="<script>"]/counter[@type="LINE"]'
 
-                        ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                        ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
+                        ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                        ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
                     }
 
                     It 'Should return the correct statistics for the method counter METHOD' {
-                        $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+                        $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
                         $result | Should -BeOfType [System.Xml.XmlDocument]
 
                         $xPath = '/report/package[@name="3.0.0"]/class/method[@name="<script>"]/counter[@type="METHOD"]'
 
-                        ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-                        ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 1
+                        ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+                        ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 1
                     }
                 }
             }
         }
 
         It 'Should return the correct statistics for the report counter INSTRUCTION' {
-            $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+            $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
             $result | Should -BeOfType [System.Xml.XmlDocument]
 
             $xPath = '/report/counter[@type="INSTRUCTION"]'
 
-            ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-            ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
-
-            # Write-IndentedXml $result
+            ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+            ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
         }
 
         It 'Should return the correct statistics for the report counter LINE' {
-            $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+            $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
             $result | Should -BeOfType [System.Xml.XmlDocument]
 
             $xPath = '/report/counter[@type="LINE"]'
 
-            ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-            ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 3
+            ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+            ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 3
         }
 
         It 'Should return the correct statistics for the report counter METHOD' {
-            $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+            $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
             $result | Should -BeOfType [System.Xml.XmlDocument]
 
             $xPath = '/report/counter[@type="METHOD"]'
 
-            ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-            ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 1
+            ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+            ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 1
         }
 
         It 'Should return the correct statistics for the report counter CLASS' {
-            $result = Update-JaCoCoStatistic -Document $mockXmlDocument
+            $result = Sampler\Update-JaCoCoStatistic -Document $mockXmlDocument
 
             $result | Should -BeOfType [System.Xml.XmlDocument]
 
             $xPath = '/report/counter[@type="CLASS"]'
 
-            ($result | Get-XmlAttributeValue -XPath $xPath).Missed | Should -Be 0
-            ($result | Get-XmlAttributeValue -XPath $xPath).Covered | Should -Be 1
+            ($result | Get-XmlAttribute -XPath $xPath).Missed | Should -Be 0
+            ($result | Get-XmlAttribute -XPath $xPath).Covered | Should -Be 1
         }
     }
 }